@@ -1,154 +1,143 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Rendering;
-using Microsoft.EntityFrameworkCore;
-using Impf_App.Data;
-using Impf_App.Models;
+﻿namespace Impf_App.Controllers;
 
-namespace Impf_App.Controllers
+public class PatientsController : Controller
 {
-    public class PatientsController : Controller
+    private readonly ApplicationDbContext _context;
+
+    public PatientsController(ApplicationDbContext context)
     {
-        private readonly ApplicationDbContext _context;
+        _context = context;
+    }
 
-        public PatientsController(ApplicationDbContext context)
+    // GET: Patients
+    public async Task<IActionResult> Index()
+    {
+        return View(await _context.Patients.ToListAsync());
+    }
+
+    // GET: Patients/Details/5
+    public async Task<IActionResult> Details(Guid? id)
+    {
+        if (id == null)
         {
-            _context = context;
+            return NotFound();
         }
 
-        // GET: Patients
-        public async Task<IActionResult> Index()
+        var patient = await _context.Patients
+            .FirstOrDefaultAsync(m => m.P_InsuranceNr == id);
+        if (patient == null)
         {
-            return View(await _context.Patients.ToListAsync());
+            return NotFound();
         }
 
-        // GET: Patients/Details/5
-        public async Task<IActionResult> Details(Guid? id)
+        return View(patient);
+    }
+
+    // GET: Patients/Create
+    public IActionResult Create()
+    {
+        return View();
+    }
+
+    // POST: Patients/Create
+    // To protect from overposting attacks, enable the specific properties you want to bind to.
+    // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
+    [HttpPost]
+    [ValidateAntiForgeryToken]
+    public async Task<IActionResult> Create([Bind("P_InsuranceNr,Sex,FirstName,LastName,BirtHDate,AdrNmbr,PLZ,Town")] Patient patient)
+    {
+        if (ModelState.IsValid)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var patient = await _context.Patients
-                .FirstOrDefaultAsync(m => m.P_InsuranceNr == id);
-            if (patient == null)
-            {
-                return NotFound();
-            }
-
-            return View(patient);
-        }
-
-        // GET: Patients/Create
-        public IActionResult Create()
-        {
-            return View();
-        }
-
-        // POST: Patients/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("P_InsuranceNr,Sex,FirstName,LastName,BirtHDate,AdrNmbr,PLZ,Town")] Patient patient)
-        {
-            if (ModelState.IsValid)
-            {
-                patient.P_InsuranceNr = Guid.NewGuid();
-                _context.Add(patient);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
-            }
-            return View(patient);
-        }
-
-        // GET: Patients/Edit/5
-        public async Task<IActionResult> Edit(Guid? id)
-        {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var patient = await _context.Patients.FindAsync(id);
-            if (patient == null)
-            {
-                return NotFound();
-            }
-            return View(patient);
-        }
-
-        // POST: Patients/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(Guid id, [Bind("P_InsuranceNr,Sex,FirstName,LastName,BirtHDate,AdrNmbr,PLZ,Town")] Patient patient)
-        {
-            if (id != patient.P_InsuranceNr)
-            {
-                return NotFound();
-            }
-
-            if (ModelState.IsValid)
-            {
-                try
-                {
-                    _context.Update(patient);
-                    await _context.SaveChangesAsync();
-                }
-                catch (DbUpdateConcurrencyException)
-                {
-                    if (!PatientExists(patient.P_InsuranceNr))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
-                }
-                return RedirectToAction(nameof(Index));
-            }
-            return View(patient);
-        }
-
-        // GET: Patients/Delete/5
-        public async Task<IActionResult> Delete(Guid? id)
-        {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var patient = await _context.Patients
-                .FirstOrDefaultAsync(m => m.P_InsuranceNr == id);
-            if (patient == null)
-            {
-                return NotFound();
-            }
-
-            return View(patient);
-        }
-
-        // POST: Patients/Delete/5
-        [HttpPost, ActionName("Delete")]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(Guid id)
-        {
-            var patient = await _context.Patients.FindAsync(id);
-            _context.Patients.Remove(patient);
+            patient.P_InsuranceNr = Guid.NewGuid();
+            _context.Add(patient);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
+        return View(patient);
+    }
 
-        private bool PatientExists(Guid id)
+    // GET: Patients/Edit/5
+    public async Task<IActionResult> Edit(Guid? id)
+    {
+        if (id == null)
         {
-            return _context.Patients.Any(e => e.P_InsuranceNr == id);
+            return NotFound();
         }
+
+        var patient = await _context.Patients.FindAsync(id);
+        if (patient == null)
+        {
+            return NotFound();
+        }
+        return View(patient);
+    }
+
+    // POST: Patients/Edit/5
+    // To protect from overposting attacks, enable the specific properties you want to bind to.
+    // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
+    [HttpPost]
+    [ValidateAntiForgeryToken]
+    public async Task<IActionResult> Edit(Guid id, [Bind("P_InsuranceNr,Sex,FirstName,LastName,BirtHDate,AdrNmbr,PLZ,Town")] Patient patient)
+    {
+        if (id != patient.P_InsuranceNr)
+        {
+            return NotFound();
+        }
+
+        if (ModelState.IsValid)
+        {
+            try
+            {
+                _context.Update(patient);
+                await _context.SaveChangesAsync();
+            }
+            catch (DbUpdateConcurrencyException)
+            {
+                if (!PatientExists(patient.P_InsuranceNr))
+                {
+                    return NotFound();
+                }
+                else
+                {
+                    throw;
+                }
+            }
+            return RedirectToAction(nameof(Index));
+        }
+        return View(patient);
+    }
+
+    // GET: Patients/Delete/5
+    public async Task<IActionResult> Delete(Guid? id)
+    {
+        if (id == null)
+        {
+            return NotFound();
+        }
+
+        var patient = await _context.Patients
+            .FirstOrDefaultAsync(m => m.P_InsuranceNr == id);
+        if (patient == null)
+        {
+            return NotFound();
+        }
+
+        return View(patient);
+    }
+
+    // POST: Patients/Delete/5
+    [HttpPost, ActionName("Delete")]
+    [ValidateAntiForgeryToken]
+    public async Task<IActionResult> DeleteConfirmed(Guid id)
+    {
+        var patient = await _context.Patients.FindAsync(id);
+        _context.Patients.Remove(patient);
+        await _context.SaveChangesAsync();
+        return RedirectToAction(nameof(Index));
+    }
+
+    private bool PatientExists(Guid id)
+    {
+        return _context.Patients.Any(e => e.P_InsuranceNr == id);
     }
 }
