@@ -1,154 +1,143 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Rendering;
-using Microsoft.EntityFrameworkCore;
-using Impf_App.Data;
-using Impf_App.Models;
+﻿namespace Impf_App.Controllers;
 
-namespace Impf_App.Controllers
+public class VaccinesController : Controller
 {
-    public class VaccinesController : Controller
+    private readonly ApplicationDbContext _context;
+
+    public VaccinesController(ApplicationDbContext context)
     {
-        private readonly ApplicationDbContext _context;
+        _context = context;
+    }
 
-        public VaccinesController(ApplicationDbContext context)
+    // GET: Vaccines
+    public async Task<IActionResult> Index()
+    {
+        return View(await _context.Vaccines.ToListAsync());
+    }
+
+    // GET: Vaccines/Details/5
+    public async Task<IActionResult> Details(Guid? id)
+    {
+        if (id == null)
         {
-            _context = context;
+            return NotFound();
         }
 
-        // GET: Vaccines
-        public async Task<IActionResult> Index()
+        var vaccine = await _context.Vaccines
+            .FirstOrDefaultAsync(m => m.P_VaccineId == id);
+        if (vaccine == null)
         {
-            return View(await _context.Vaccines.ToListAsync());
+            return NotFound();
         }
 
-        // GET: Vaccines/Details/5
-        public async Task<IActionResult> Details(Guid? id)
+        return View(vaccine);
+    }
+
+    // GET: Vaccines/Create
+    public IActionResult Create()
+    {
+        return View();
+    }
+
+    // POST: Vaccines/Create
+    // To protect from overposting attacks, enable the specific properties you want to bind to.
+    // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
+    [HttpPost]
+    [ValidateAntiForgeryToken]
+    public async Task<IActionResult> Create([Bind("P_VaccineId,Description,Manufacturer,Dosis,DateOfApproval")] Vaccine vaccine)
+    {
+        if (ModelState.IsValid)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var vaccine = await _context.Vaccines
-                .FirstOrDefaultAsync(m => m.P_VaccineId == id);
-            if (vaccine == null)
-            {
-                return NotFound();
-            }
-
-            return View(vaccine);
-        }
-
-        // GET: Vaccines/Create
-        public IActionResult Create()
-        {
-            return View();
-        }
-
-        // POST: Vaccines/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("P_VaccineId,Description,Manufacturer,Dosis,DateOfApproval")] Vaccine vaccine)
-        {
-            if (ModelState.IsValid)
-            {
-                vaccine.P_VaccineId = Guid.NewGuid();
-                _context.Add(vaccine);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
-            }
-            return View(vaccine);
-        }
-
-        // GET: Vaccines/Edit/5
-        public async Task<IActionResult> Edit(Guid? id)
-        {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var vaccine = await _context.Vaccines.FindAsync(id);
-            if (vaccine == null)
-            {
-                return NotFound();
-            }
-            return View(vaccine);
-        }
-
-        // POST: Vaccines/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(Guid id, [Bind("P_VaccineId,Description,Manufacturer,Dosis,DateOfApproval")] Vaccine vaccine)
-        {
-            if (id != vaccine.P_VaccineId)
-            {
-                return NotFound();
-            }
-
-            if (ModelState.IsValid)
-            {
-                try
-                {
-                    _context.Update(vaccine);
-                    await _context.SaveChangesAsync();
-                }
-                catch (DbUpdateConcurrencyException)
-                {
-                    if (!VaccineExists(vaccine.P_VaccineId))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
-                }
-                return RedirectToAction(nameof(Index));
-            }
-            return View(vaccine);
-        }
-
-        // GET: Vaccines/Delete/5
-        public async Task<IActionResult> Delete(Guid? id)
-        {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var vaccine = await _context.Vaccines
-                .FirstOrDefaultAsync(m => m.P_VaccineId == id);
-            if (vaccine == null)
-            {
-                return NotFound();
-            }
-
-            return View(vaccine);
-        }
-
-        // POST: Vaccines/Delete/5
-        [HttpPost, ActionName("Delete")]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(Guid id)
-        {
-            var vaccine = await _context.Vaccines.FindAsync(id);
-            _context.Vaccines.Remove(vaccine);
+            vaccine.P_VaccineId = Guid.NewGuid();
+            _context.Add(vaccine);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
+        return View(vaccine);
+    }
 
-        private bool VaccineExists(Guid id)
+    // GET: Vaccines/Edit/5
+    public async Task<IActionResult> Edit(Guid? id)
+    {
+        if (id == null)
         {
-            return _context.Vaccines.Any(e => e.P_VaccineId == id);
+            return NotFound();
         }
+
+        var vaccine = await _context.Vaccines.FindAsync(id);
+        if (vaccine == null)
+        {
+            return NotFound();
+        }
+        return View(vaccine);
+    }
+
+    // POST: Vaccines/Edit/5
+    // To protect from overposting attacks, enable the specific properties you want to bind to.
+    // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
+    [HttpPost]
+    [ValidateAntiForgeryToken]
+    public async Task<IActionResult> Edit(Guid id, [Bind("P_VaccineId,Description,Manufacturer,Dosis,DateOfApproval")] Vaccine vaccine)
+    {
+        if (id != vaccine.P_VaccineId)
+        {
+            return NotFound();
+        }
+
+        if (ModelState.IsValid)
+        {
+            try
+            {
+                _context.Update(vaccine);
+                await _context.SaveChangesAsync();
+            }
+            catch (DbUpdateConcurrencyException)
+            {
+                if (!VaccineExists(vaccine.P_VaccineId))
+                {
+                    return NotFound();
+                }
+                else
+                {
+                    throw;
+                }
+            }
+            return RedirectToAction(nameof(Index));
+        }
+        return View(vaccine);
+    }
+
+    // GET: Vaccines/Delete/5
+    public async Task<IActionResult> Delete(Guid? id)
+    {
+        if (id == null)
+        {
+            return NotFound();
+        }
+
+        var vaccine = await _context.Vaccines
+            .FirstOrDefaultAsync(m => m.P_VaccineId == id);
+        if (vaccine == null)
+        {
+            return NotFound();
+        }
+
+        return View(vaccine);
+    }
+
+    // POST: Vaccines/Delete/5
+    [HttpPost, ActionName("Delete")]
+    [ValidateAntiForgeryToken]
+    public async Task<IActionResult> DeleteConfirmed(Guid id)
+    {
+        var vaccine = await _context.Vaccines.FindAsync(id);
+        _context.Vaccines.Remove(vaccine);
+        await _context.SaveChangesAsync();
+        return RedirectToAction(nameof(Index));
+    }
+
+    private bool VaccineExists(Guid id)
+    {
+        return _context.Vaccines.Any(e => e.P_VaccineId == id);
     }
 }
